requirements.txt
Django==1.7.5
django-sekizai
django-tastypie